+++ src/components/organisms/about/index.js
@@ -1,0 +1,1 @@
+export * from './about.jsx'
