+++ src/components/organisms/education/index.js
@@ -1,0 +1,1 @@
+export * from './education.jsx'
