+++ src/components/organisms/experiences/index.js
@@ -1,0 +1,1 @@
+export * from "./experiences.jsx";
