+++ src/components/organisms/general/index.js
@@ -1,0 +1,1 @@
+export * from "./general.jsx";
