+++ src/components/organisms/index.js
@@ -1,0 +1,7 @@
+export * from "./about";
+export * from "./education";
+export * from "./experiences";
+export * from "./interests";
+export * from "./projects";
+export * from "./skills";
+export * from "./general";
