+++ src/components/organisms/interests/index.js
@@ -1,0 +1,1 @@
+export * from "./interests.jsx";
