+++ src/components/organisms/projects/index.js
@@ -1,0 +1,1 @@
+export * from "./projects.jsx";
