+++ src/components/organisms/skills/index.js
@@ -1,0 +1,1 @@
+export * from "./skills.jsx";
