+++ src/components/pages/index.js
@@ -1,0 +1,1 @@
+export * from "./main";
