+++ src/components/pages/main/index.js
@@ -1,0 +1,1 @@
+export * from "./main.jsx";
